requirements.txt
requests~=2.32.4
pytest~=8.4.1
Faker~=37.5.3
python-dotenv~=1.1.1